backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
ultralytics==8.1.0  # YOLO v8
deepface==0.0.87
numpy==1.24.3
Pillow==10.2.0

# Tracking
filterpy==1.4.5  # Kalman filter for tracking

# Audio Processing
openai-whisper==20231117
pyannote.audio==3.1.1
resemblyzer==0.1.1.dev0

# NLP & Sentiment
transformers==4.37.2
torch==2.1.2
vaderSentiment==3.3.2
sentencepiece==0.1.99

# ONNX Runtime for FastVLM
onnxruntime==1.17.0
onnx==1.15.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
